deleted some comments, made the accelerometer more sensitive
Former-commit-id: 0bf5064ac439a8f4f74431fb92b9848d6c4add3e [formerly 15429df48c620910c9df72ac93cdad11f688230a]
Former-commit-id: e4712745d323715af848d0c772ba3116a3c1aa44

+++ catasstrophy/Countdown.m
@@ -141,7 +141,7 @@
         SKScene * game = [[MyScene alloc] initWithSize:self.size];
         [self.view presentScene:game transition:[SKTransition fadeWithDuration:0]];
     }else {
-        self.countdownLabel.text = [NSString stringWithFormat:@"%i", self.count];
+        self.countdownLabel.text = [NSString stringWithFormat:@"%li", (long)self.count];
     }
     
     [self.countdownLabel runAction:[SKAction sequence:@[unfadeCountdown, fadeCountdown]]];
